requirements.txt
langchain==0.3.17
langgraph==0.2.68
langchain-community==0.3.16
langchain-openai==0.3.2
databricks_langchain==0.2.0
faiss-cpu==1.9.0
grandalf==0.8
langsmith==0.3.5
bs4==0.0.2
